variable expansion of package names
git-svn-id: https://svn.iu.hio.no/projects/cfengine-2/trunk@465 0bacebff-10ef-0310-a86f-bc4055e020e4

+++ src/image.c
@@ -1514,8 +1514,15 @@ if (!FixCompressedArrayValue(i,value,&(ip->inode_cache)))
       }
    else
       {
-      Verbose("Removing old hard link %s to preserve structure..\n",value);
-      unlink(value);
+      if (strcmp(ip->action,"warn") == 0)
+          {
+          Verbose("Need to remove old hard link %s to preserve structure..\n",value);
+          }
+      else
+          {
+          Verbose("Removing old hard link %s to preserve structure..\n",value);
+          unlink(value);
+          }
       }
    }
 }
